Fix iterating using only a timestamp column
When the cursor is created with only one time column, decoding it would
crash, reporting that the columns and values have different sizes.

This is because `Array(values)` attempts to call `#to_ary` and `#to_a`
on its arguments before manually wrapping it in an array. This means
that `Time#to_a` is called, which returns an array of 10 elements.